changed api
